requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
openai==1.30.1
python-docx==1.1.2
PyPDF2==3.0.1
python-multipart==0.0.9
